+++ src/Effects/PlayerEffects/SetMaxStatEffect.cs
@@ -37,28 +37,26 @@ public sealed class SetMaxStatEffect : CrowdControlEffect
     protected override CrowdControlResponseStatus OnStart()
     {
         var player = GetLocalPlayer();
-        if ((_life && _increase && player.Player.statLifeMax >= 500) ||
-            (_life && !_increase && player.Player.statLifeMax <= 20) ||
-            (!_life && _increase && player.Player.statManaMax >= 200) ||
-            (!_life && !_increase && player.Player.statManaMax <= 20))
-        {
-            // Ignore if the stat cannot be further altered
-            return CrowdControlResponseStatus.Failure;
-        }
-
         if (_life)
         {
             if (_increase)
             {
                 // Increase the player's current max health
-                player.Player.statLifeMax += Amount;
+                if (player.Player.AddStatLifeMax(Amount) == Amount)
+                {
+                    return CrowdControlResponseStatus.Failure;
+                }
+
                 player.Player.statLife += Amount;
                 player.Player.AddBuff(BuffID.Lovestruck, 60 * 5);
             }
             else
             {
                 // Decrease the player's current max health
-                player.Player.statLifeMax -= Amount;
+                if (player.Player.AddStatLifeMax(-Amount) == -Amount)
+                {
+                    return CrowdControlResponseStatus.Failure;
+                }
             }
         }
         else
@@ -66,13 +64,20 @@ public sealed class SetMaxStatEffect : CrowdControlEffect
             if (_increase)
             {
                 // Increase the player's current max mana
-                player.Player.statManaMax += Amount;
+                if (player.Player.AddStatManaMax(Amount) == Amount)
+                {
+                    return CrowdControlResponseStatus.Failure;
+                }
+
                 player.Player.statMana += Amount;
             }
             else
             {
                 // Decrease the player's current max mana
-                player.Player.statManaMax -= Amount;
+                if (player.Player.AddStatManaMax(-Amount) == -Amount)
+                {
+                    return CrowdControlResponseStatus.Failure;
+                }
             }
         }